SciGlassDatabase.SciGlassDatabase — Module

A Julia package for extracting data from SciGlass database.

Verbatim from https://github.com/epam/SciGlass

SciGlass

The largest glass property database contains data for more than 420 thousand glass compositions including more than 18 thousand halide and about 38 thousand chalcogenide glasses. It provides also property predictions and calculations, help you solve R&D problems.

Features

Glass Properties. Practically all broadly used physical and chemical properties of glasses and glass-forming melts with concise but informative description of syntheses and measurement procedures.

Glasses. 422,000 glasses and melts, including more than 268,000 oxide glasses and melts, 18,500 halide and 38,500 chalcogenide glasses. The data were taken from more than 40,000 literature sources including more than 19,700 patents.

Property Calculations. Over 100 computational methods to compute the properties in 15 groups (e.g., viscosity, density, mechanical, optical), many of them in broad temperature ranges. Prediction of properties of oxide, halide, and chalcogenide glasses is possible in wide concentration ranges.

Ternary Diagrams of Glass Formation. More than 3,800 ternary diagrams of glass formation.

Optical Spectra. More than 15,000 optical spectra (from UV to near IR) for glasses and melts with 96 different ways to represent spectra.

Ternary Property Diagram.Automatically generate isoproperty lines versus composition and compare calculated and experimental property values for ternary compositions.

Statistical Analysis. Find best computational method from least squares fit of calculated and experimental values to insure best method can be applied to your glass composition.

Patent and Trademark Index. Over 19,000 international patents and 1,000 trademarks; complete information on usage, country, company, composition, property table, author, and TM symbol.

Subject Index. Explore hundreds of specialized subjects (e.g., diffusion of specific gases & ions), which are difficult to find by other ways.

Chemical Durability of Glasses. Pertinent data on about 35,000 glasses as well as access to a large compendium on standard durability testing methods.

Optimization of Glass Compositions. Find the most perspective glass compositions meeting a complex of requirements on specific values of their properties

SciGlassDatabase.Property — Type

Property holds all the table in Property.mdb database file.

List of tables:

ATOMS
CalcListProp
IDF
Legends
SYMBOL
tblpar
CalcAtoms
DT
Journal
PConvert
Sort_Comp
CalcComp
DT2
LISTPROP
PropList
UNITS

SciGlassDatabase.Select — Type

Select holds all the table in Select.mdb database file.

List of tables:

Authors
GF_add
MaxGno
Reference
SpectralParam
WtPc
AtMol
Contries
Gcomp
MolPc
SciGK
SubjectIndex
AtWt
GF
Kod2Ref
Patents
Spectr
Trademark

SciGlassDatabase.clean_df! — Method

clean_df!(df, item::Symbol)

Cleans and preprocesses a DataFrame loaded from the SciGlass database. This function applies specific cleaning rules based on the type of table being processed.

Arguments

df: A DataFrame object representing the data to be cleaned.

item: A Symbol representing the name of the table in the database (e.g., :SciGK).

Description

The clean_df! function modifies the input DataFrame in place, applying transformations such as renaming columns or handling inconsistencies. Each table type may have its own cleaning rules, which are determined by the item argument.

SciGlassDatabase.eachtable — Method

eachtable(tables::Union{Property, Select})

Generate iterator over tables in Union{Property, Select}.

Arguments

tables: A collection of tables. Each table should be iterable and have a names function that returns its column names.

Example

for (name, table) in eachtable(SELECT)
    # some code
end

SciGlassDatabase.get_components_from_composition — Method

get_components_from_composition(x; molp=true)

Return list of components with data.

SciGlassDatabase.get_composition_from_glass_number — Method

get_composition_from_glass_number(glassno; composition="SiO2", molp=true, inner=true)

Return mask if components present in glass numbers and glass compositions from glass numbers.

SciGlassDatabase.get_compositions — Method

get_compositions(;property=["DENSITY", "TLiq"], composition="SiO2-K2O", prop_inner=true, comp_inner=true,
    clean=true, molp=true, with_metadata=false)

Return a DataFrame with glass compositions and their properties.

Arguments

property=["DENSITY"]: List of properties to be extracted.

composition="SiO2-K2O": Glass compositions.

prop_inner=true: If true takes intersection of data for multiple properties, otherwise union.

comp_inner=true: If true takes intersection of data for multiple components, otherwise union.

clean=true": If true, clean the DataFrame before returning (extra columns removed).

molp=true: If true, compositions are in mol%, otherwise weight%.

with_metadata=false: If true, give metadata for each composition (Authors, Year, Journal etc.).

Returns

DataFrame: DataFrame with glass compositions and their properties.

DataFrame contains the following columns:

GLASNO: Glass number.

property: Property values.

composition: Composition values.

Example

get_compositions(;property=["DENSITY", "TLiq"], composition="SiO2-Na2O")

SciGlassDatabase.get_molar_wt — Method

get_molar_wt(comps_list)

Return molar weight from component list.

Arguments

comps_list: List of glass components

Returns

'list': List of molar weight for each component

SciGlassDatabase.list_metadata — Method

list_metadata()

Return vector of subset of metadata headers for a glass compositions.

Returns

list: List of metadata headers.

List of metadata:

Author, Year, GlassClass, Analysis, PropCode, GForm, any_n, Trademark, QComp, QComp2, Flag

SciGlassDatabase.list_properties — Method

list_properties()

Return property names as vector that can be extracted using get_compositions.

Returns

list: List of property names.

List of properties:

T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, V500, V600, V700, V800, V900, V1000, V1100, V1200, V1300, V1400, V1500, V1600, V1800, V2000, V2200, TG, LPT, ANPT, SPT, Tsoft, RO150, RO300, RO20, RO100, TK100C, TEC55, TEC100, TEC160, TEC210, TEC350, ANYTEC, DENSITY, spvm, ND300, nd300low, nd300hi, DNFC300, NUD300, RTSH180, MODUNG, MODSDV, MIKROTV, EPS730, TGD730, ro800, ro1000, ro1200, ro1400, Tm, TLiq, SUT900, SUT1200, SUT1300, SUT1400, anysut, cond220, pois, cp20, cp200, cp400, cp800, cp1000, cp1200, cp1400, dens800, dens1000, dens1200, dens1400, Tmax, Vmax, Tn, Io, tcr, tx

SciGlassDatabase.load_all_tables! — Method

load_all_tables!(tables; force=false)

Load all tables in the given collection of tables. If force is set to true, it will reload the tables even if they are already loaded.

Arguments

tables: A collection of tables. Each table should be iterable and have a names function that returns its column names.

Keyword Arguments

force: A boolean flag indicating whether to force reload the tables. Default is false.

See also reload_table!, load_table!, load_all_tables!.

SciGlassDatabase.load_table! — Method

load_table!(table, item::Union{Symbol, String})

Set table field with DataFrame from filepath. Skip loading if DataFrame already exists, use reload_table! for force reloading of table. If the filepath is a string, it will be joined with the artifact path.

Arguments

table: A collection of tables. Each table should be iterable and have a names function that returns its column names.

item: The name of the table to load. It can be a Symbol or String.

Example

load_table!(SELECT, "AtMol")

See also reload_table!, load_table!, load_all_tables!.

SciGlassDatabase.print_all_headers! — Method

print_all_headers!(tables)

Prints the headers (column names) of all tables in the given collection of tables.

Arguments

tables: A collection of tables. Each table should be iterable and have a names function that returns its column names.

Behavior

Iterates through each table in the collection.

Prints the name of the table (key) in bold red text.

Prints the column names of the table.

If an error occurs while processing a table, it catches the exception and prints a message indicating the table was skipped.

SciGlassDatabase.print_properties — Method

print_properties()

Print property names from the list_properties.

SciGlassDatabase.print_property_data — Method

print_property_data()

Print property names and their corresponding field names from the PROPERTY.LISTPROP table.

SciGlassDatabase.reload_table! — Method

reload_table!(tables, item::Union{Symbol, String})

Reload the table field with DataFrame from filepath. Skip loading if DataFrame already exists, use reload_table! for force reloading of table.

Arguments

tables: A collection of tables. Each table should be iterable and have a names function that returns its column names.

item: The name of the table to load. It can be a Symbol or String.

Example

reload_table!(SELECT, "AtMol")

SciGlassDatabase.save_table — Method

save_table(filepath, df)

Saves a DataFrame to a CSV file at the specified filepath.

Arguments

filepath: A String specifying the path where the CSV file will be saved.

df: A DataFrame object to be saved.

Description

This function writes the contents of the DataFrame to a CSV file using the CSV.write function.

SciGlassDatabase.scatterTernary — Method

scatterTernary(composition; property="DENSITY", kwargs...)

Generates a scatter plot on a ternary diagram based on the given composition and property.

Arguments

composition: The composition data to be plotted. This should be a data structure compatible with the get_compositions function.

property: A string specifying the property to be visualized. Defaults to "DENSITY". The property name is automatically converted to uppercase.

kwargs...: Additional keyword arguments passed to the get_compositions function.

Returns

A scatter plot object if the composition data contains entries; otherwise, no plot is generated.

Details

The function retrieves the composition data and the specified property using the get_compositions function.

The ternary composition data is converted to Cartesian coordinates using the tern2cart function.

A ternary plot is created with axes labeled according to the composition names.

The scatter plot is colored based on the values of the specified property, with a colorbar included for reference.

Notes

The function assumes that the input composition data contains at least three components for ternary plotting.

If the composition data is empty, no scatter plot is generated.

SciGlassDatabase.scatterplot_table — Method

scatterplot_table(df, x, y)

Generates a scatter plot from the given DataFrame df using the specified columns x and y.

Arguments

df::DataFrame: The input DataFrame containing the data to be plotted.

x::Symbol: The column name in the DataFrame to be used for the x-axis.

y::Symbol: The column name in the DataFrame to be used for the y-axis.

Behavior

Converts the column names x and y to uppercase before plotting.

Creates a scatter plot with no legend label.

Sets the x-axis and y-axis labels to the uppercase string representation of the column names.

SciGlassDatabase.separate_columns — Method

separate_columns(comps, compositions)

Convert composition from string representation to DataFrame.

SciGlassDatabase.set_table_field! — Method

set_table_field!(a, item::Symbol, filepath::DataFrame)

Set table field with DataFrame from filepath. Skip loading if DataFrame already exists, use reload_table! for force reloading of table.

Arguments

a: A collection of tables. Each table should be iterable and have a names function that returns its column names.

item: The name of the table to load. It can be a Symbol or String.

filepath: The path to the file containing the table data. It can be a String or DataFrame.

SciGlassDatabase.set_table_field! — Method

set_table_field!(a, item::Symbol, filepath::String)

Set table field with DataFrame from filepath. Skip loading if DataFrame already exists, use reload_table! for force reloading of table.

Arguments

a: A collection of tables. Each table should be iterable and have a names function that returns its column names.

item: The name of the table to load. It can be a Symbol or String.

filepath: The path to the file containing the table data. It can be a String or DataFrame.

SciGlassDatabase.show_table — Method

show_table(df; title="Table", with_plot=false)

Displays a table using Dash and optionally includes an interactive plot.

Arguments

df::DataFrame: The data to be displayed in the table.

title::String: The title of the table. Defaults to "Table".

with_plot::Bool: If true, includes an interactive plot feature. Defaults to false.

Description

This function creates a Dash application to display a table. If with_plot is set to true, it also provides an interactive scatter plot feature using PlotlyJS. Users can select the columns for the x and y axes of the plot through dropdown menus.

Interactive Plot

When with_plot is enabled:

Dropdown menus allow users to select columns for the x and y axes.

A scatter plot is generated dynamically based on the selected columns.

The plot is displayed alongside the table.

SciGlassDatabase.table_layout! — Method

table_layout!(app, df; title="Table", with_plot=false)

Sets up the layout for the Dash application.

Arguments

app: The Dash application instance.

df: A DataFrame object to be displayed in the table.

title: A String specifying the title of the table. Defaults to "Table".

with_plot: A Boolean indicating whether to include a plot in the layout. Defaults to false.

Description

This function defines the layout of the Dash application, including the table, optional plot, and dropdowns for selecting x and y axes for the plot.